refactor(checl): move to subdir
